truetribe-backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
Pillow
redis
celery
psycopg2-binary
python-decouple
cloudinary
opencv-python
face-recognition
scikit-learn
numpy
pandas
channels
channels-redis
daphne
django-extensions
faker
pytest
pytest-django
